divded py file
calculator function divide from main.py
