+++ calcu_function.py
@@ -1,0 +1,15 @@
+# This function adds two numbers
+def add(x, y):
+    return x + y
+
+# This function subtracts two numbers
+def subtract(x, y):
+    return x - y
+
+# This function multiplies two numbers
+def multiply(x, y):
+    return x * y
+
+#Need to define divide function.
+def divide (x,y):
+    return x/y
